Add files via upload
The fill-in-the-blanks quiz is almost feature complete and bug free.  There's a fill-in-the-blanks.pyc file in here which is a compiled version of this project to demonstrate how the project should function.  The file I created is the fill-in-the-blanks.py file.  I also created a file called answers.txt to accompany the quiz; when someone submits an incorrect answer, it logs the answer and relevant details for training and troubleshooting, plus it catches a few Unexpected Results showing some details that could help diagnose if things don't function as expected.

The instructions mention the ability to type in "PASS" to skip that blank and come back to it later, but I will build that feature next, ignore for now.

Also, I am troubleshooting a bug that doesn't recognize some of the blanks even when they are formatted exactly the same as other blanks that are recognized (causing the program to declare victory prematurely).  My temporary work around are to put a space before and after t